解决后台服务数据悄悄修改数据类型，导致前端数据类型解析报错的问题

+++ app/src/main/java/com/yyide/chatim/activity/leave/AddressBookActivity.java
@@ -280,7 +280,7 @@ public class AddressBookActivity extends BaseMvpActivity<AddressBookPresenter> i
 
     private List<String> getIds(List<AddressBookBean> noticeScopeBeans, List<String> ids) {
         for (AddressBookBean noticeScopeBean : noticeScopeBeans) {
-            int id = noticeScopeBean.getId();
+            long id = noticeScopeBean.getId();
             ids.add("" + id);
             List<AddressBookBean> list = noticeScopeBean.getList();
             if (list != null && !list.isEmpty()) {
